pyproject.toml

[project]
name = "flask-devkit"
version = "0.1.0"
packages = [{include = "flask_devkit", from = ""}]
description = "A comprehensive Flask extension for rapid, secure API development with built-in JWT authentication and RBAC."
authors = [
    {name = "Bahaa Shklawon", email = "bahaashklawon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "MIT"}
keywords = ["flask", "auth", "authentication", "rbac", "jwt", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Framework :: Flask",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "apiflask (>=2.4.0,<3.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-limiter (>=3.5.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "flask_devkit", from = ""}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^5.0.0"
alembic = "^1.13.2"
ruff = "^0.5.7"
black = "^24.8.0"
mypy = "^1.11.2"
pre-commit = "^3.8.0"

[tool.pytest.ini_options]
addopts = "-q --cov=flask_devkit --cov-report=term-missing --cov-fail-under=80"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"]
exclude = ["migrations/*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = false
